+++ src/Commands/InstallWildfireCommand.php
@@ -25,6 +25,16 @@ class InstallWildfireCommand extends InstallCommand
     protected $library = 'wildfire';
 
     /**
+     * Checks whether the command is enabled or not in the current environment.
+     *
+     * @return boolean
+     */
+    public function isEnabled()
+    {
+        return ! file_exists(APPPATH . 'libraries/Doctrine.php');
+    }
+
+    /**
      * Executes the command.
      * 
      * @param  \Symfony\Component\Console\Input\InputInterface   $input